+++ app/cpt_file/soil_layout_conversion_functions.py
@@ -15,21 +15,24 @@ CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFT
 SOFTWARE.
 """
 from copy import deepcopy
+from io import BytesIO, StringIO
 from math import ceil
-from typing import List
+from typing import List, Union
 
-from viktor import Color
-from viktor import UserException
-from viktor.geo import GEFClassificationError
-from viktor.geo import GEFFile
-from viktor.geo import GEFParsingException
-from viktor.geo import RobertsonMethod
-from viktor.geo import Soil
-from viktor.geo import SoilLayer
-from viktor.geo import SoilLayout
-from .constants import ADDITIONAL_COLUMNS
-from .constants import DEFAULT_MIN_LAYER_THICKNESS
-from .constants import DEFAULT_ROBERTSON_TABLE
+from munch import Munch, unmunchify
+from viktor import Color, UserException
+from viktor.geo import (
+    GEFClassificationError,
+    GEFFile,
+    GEFParsingException,
+    RobertsonMethod,
+    Soil,
+    SoilLayer,
+    SoilLayout,
+    TableMethod,
+)
+
+from .constants import ADDITIONAL_COLUMNS, DEFAULT_MIN_LAYER_THICKNESS
 
 
 def convert_soil_layout_from_mm_to_meter(soil_layout: SoilLayout) -> SoilLayout:
@@ -53,92 +56,178 @@ def convert_soil_layout_from_meter_to_mm(soil_layout: SoilLayout) -> SoilLayout:
 def filter_nones_from_params_dict(raw_dict) -> dict:
     """Removes all rows which contain one or more None-values"""
     rows_to_be_removed = []
-    for row_index, items in enumerate(zip(*raw_dict['measurement_data'].values())):
+    for row_index, items in enumerate(zip(*raw_dict["measurement_data"].values())):
         if None in items:
             rows_to_be_removed.append(row_index)
     for row in reversed(rows_to_be_removed):
-        for signal in raw_dict['measurement_data'].keys():
-            del raw_dict['measurement_data'][signal][row]
+        for signal in raw_dict["measurement_data"].keys():
+            del raw_dict["measurement_data"][signal][row]
     return raw_dict
 
 
-def convert_input_table_field_to_soil_layout(bottom_of_soil_layout_user: float,
-                                             soil_layers_from_table_input: List[dict]) -> SoilLayout:
-    """Creates a viktor SoilLayout from the user input."""
-    bottom = bottom_of_soil_layout_user
-    soils = get_soil_mapping()
-    soil_layers = []
+def convert_input_table_field_to_soil_layout(
+    bottom_of_soil_layout_user: float,
+    soil_layers_from_table_input: List[dict],
+    soils: dict,
+) -> SoilLayout:
+    """Creates a SoilLayout from the user input.
 
+    :param bottom_of_soil_layout_user: Bottom of soil layout in [m]
+    :param soil_layers_from_table_input: Table where a row represents a layer.
+    Each row should contain a soil name and top of layer [m].
+    :param soils: Dictionary with soil names and their respective Soil.
+    :return: SoilLayout
+    """
+    bottom = bottom_of_soil_layout_user
+    soil_layers = []
     for layer in reversed(soil_layers_from_table_input):
         soil_name = layer["name"]
         top_of_layer = layer["top_of_layer"]
         try:
             soil_layers.append(SoilLayer(soils[soil_name], top_of_layer, bottom))
         except KeyError:
-            raise UserException(f"{soil_name} is not available in the selected classification " f"table.\n "
-                                f"Please select a different table, or reclassify the CPT files")
+            raise UserException(
+                f"'{soil_name}' is not available in the selected classification table.\n"
+                f"Please select a different table, or reclassify the CPT files"
+            )
         bottom = top_of_layer  # Set bottom of next soil layer to top of current layer.
-
     return convert_soil_layout_from_meter_to_mm(SoilLayout(soil_layers[::-1]))
 
 
 def convert_soil_layout_to_input_table_field(soil_layout: SoilLayout) -> List[dict]:
     """Converts a SoilLayout to the parametrisation representation (Field = InputTable)."""
     return [
-        {
-            "name": layer.soil.properties.ui_name,
-            "top_of_layer": layer.top_of_layer
-        }
-        for layer in soil_layout.layers
+        {"name": layer.soil.properties.ui_name, "top_of_layer": layer.top_of_layer} for layer in soil_layout.layers
     ]
 
 
-def get_soil_mapping() -> dict:
-    """Returns a mapping between the soil name visible in the UI and the Soil object used in the logic"""
-    soil_mapping = {}
-    for soil in DEFAULT_ROBERTSON_TABLE:
-        # remove soil color from the properties
-        properties = deepcopy(soil)
-        del properties['color']
-        # create soil mapping
-        soil_mapping[soil['ui_name']] = Soil(soil['name'], Color(*soil['color']), properties=properties)
-    return soil_mapping
+def convert_to_color(rgb: Union[str, tuple]) -> Color:
+    """Simple conversion function that always returns a Color object"""
+    if isinstance(rgb, tuple):
+        return Color(*rgb)
+    return Color(*[int(element) for element in rgb.strip().split(",")])
 
 
-def classify_cpt_file_on_upload(cpt_file: GEFFile) -> dict:
-    """Classify an uploaded CPT File based on the selected _ClassificationMethod"""
+def get_water_level(cpt_data_object) -> float:
+    """Water level is assigned to the value parsed from GEF file if it exists, otherwise a default is assigned to 1m
+    below the surface level"""
+    if hasattr(cpt_data_object, "water_level"):
+        water_level = cpt_data_object.water_level
+    else:
+        water_level = cpt_data_object.ground_level_wrt_reference / 1e3 - 1
+    return water_level
 
-    try:
-        # Parse the GEF file content
-        cpt_data_object = cpt_file.parse(additional_columns=ADDITIONAL_COLUMNS, return_gef_data_obj=True)
-        # Parse water level from GEF file if it exists
-        if hasattr(cpt_data_object, 'water_level'):
-            ground_water_level = cpt_data_object.water_level
-        else:  # a default is assigned (1m below the surface level)
-            ground_water_level = round(cpt_data_object.ground_level_wrt_reference/1e3 - 1, 2)
-        # Classify the CPTData object to get a SoilLayout
-        soil_layout_obj = cpt_data_object.classify(method=RobertsonMethod(DEFAULT_ROBERTSON_TABLE),
-                                                   return_soil_layout_obj=True)
 
-    except GEFParsingException as parsing_exception:
-        raise UserException(f"CPT Parsing: {str(parsing_exception)}") from parsing_exception
-    except GEFClassificationError as classification_exception:
-        raise UserException(f"CPT Classification: {str(classification_exception)}") from classification_exception
+def _update_classification_table(classification_table: List[dict]) -> List[dict]:
+    """Updates the table so that the min and max parameters for gamma are defined."""
+    for row in classification_table:
+        row["gamma_dry_min"] = None
+        row["gamma_dry_max"] = row["gamma_dry"]
+        row["gamma_wet_min"] = None
+        row["gamma_wet_max"] = row["gamma_wet"]
+        row["ui_name"] = row["name"]
+    return classification_table
 
-    # filter thickness and convert to meter
-    soil_layout_filtered = soil_layout_obj.filter_layers_on_thickness(
-        min_layer_thickness=DEFAULT_MIN_LAYER_THICKNESS,
-        merge_adjacent_same_soil_layers=True
-    )
-    soil_layout_filtered_in_m = convert_soil_layout_from_mm_to_meter(soil_layout_filtered)
 
-    # Serialize the parsed CPT File content and update it with the new soil layout
-    cpt_dict = cpt_data_object.serialize()
-    cpt_dict['ground_level'] = cpt_data_object.ground_level_wrt_reference/1e3
-    cpt_dict['soil_layout_original'] = soil_layout_obj.serialize()
-    cpt_dict['bottom_of_soil_layout_user'] = ceil(soil_layout_obj.bottom) / 1e3
-    cpt_dict['soil_layout'] = convert_soil_layout_to_input_table_field(soil_layout_filtered_in_m)
-    cpt_dict['ground_water_level'] = ground_water_level
-    cpt_dict['x_rd'] = cpt_dict['headers']['x_y_coordinates'][0] if 'x_y_coordinates' in cpt_dict['headers'] else 0
-    cpt_dict['y_rd'] = cpt_dict['headers']['x_y_coordinates'][1] if 'x_y_coordinates' in cpt_dict['headers'] else 0
-    return cpt_dict
+def _update_color_string(classification_table: List[dict]) -> List[dict]:
+    """Converts the RGB color strings in the table into a tuple (R, G, B)"""
+    for row in classification_table:
+        if not isinstance(row["color"], Color):
+            row["color"] = convert_to_color(row["color"])
+    return classification_table
+
+
+class Classification:
+    """This class handles all logic related to selecting the correct method and table for classification of CPTData.
+
+    It also provides the correct soil mapping needs for the visualizations of the soil layers.
+    """
+
+    def __init__(self, classification_params: Munch):
+        self._method = classification_params.method
+        self._table = unmunchify(classification_params.get(self._method))
+
+    @property
+    def table(self) -> List[dict]:
+        """Returns a cleaned up table that can be used for the Classification methods"""
+        if self._method == "robertson":
+            return _update_color_string(self._table)
+        return _update_classification_table(self._table)
+
+    def method(self, ground_water_level) -> Union[TableMethod, RobertsonMethod]:
+        """Returns the appropriate _ClassificationMethod for the CPTData.classify() function"""
+        if self._method == "robertson":
+            return RobertsonMethod(self.table)
+        if self._method == "table":
+            return TableMethod(self.table, ground_water_level=ground_water_level)
+        raise UserException(f"The {self._method} method has not yet been implemented")
+
+    def get_table_plot(self, gwl: float = 0, file_format: str = "pdf") -> Union[BytesIO, StringIO]:
+        """Returns a plot of the selected _ClassificationMethod
+        The ground water level is irrelevant for qualification plot and therefore set to zero
+        to make sure the qualification plot is always downloadable."""
+        if self._method == "robertson":
+            raise TypeError
+        if self._method == "table":
+            return self.method(gwl).get_qualification_table_plot(fileformat=file_format)
+        raise UserException(f"The {self._method} method has not yet been implemented")
+
+    def get_table_plot_svg(self, gwl: float = 0, file_format: str = "svg") -> Union[BytesIO, StringIO]:
+        """Returns a plot of the selected _ClassificationMethod"""
+        if self._method == "robertson":
+            raise TypeError
+        if self._method == "table":
+            return self.method(gwl).get_qualification_table_plot(fileformat=file_format)
+        raise UserException(f"The {self._method} method has not yet been implemented")
+
+    @property
+    def soil_mapping(self) -> dict:
+        """Returns a mapping between the soil name visible in the UI and the Soil object used in the logic"""
+        soil_mapping = {}
+        for soil in self.table:
+            ui_name = soil["ui_name"]
+            properties = deepcopy(soil)
+            if self._method == "robertson":
+                del properties["color"]
+            soil_mapping[ui_name] = Soil(soil["name"], convert_to_color(soil["color"]), properties=properties)
+        return soil_mapping
+
+    def classify_cpt_file(self, cpt_file: GEFFile, saved_ground_water_level=None) -> dict:
+        """Classify an uploaded CPT File based on the selected _ClassificationMethod"""
+
+        try:
+            # Parse the GEF file content
+            cpt_data_object = cpt_file.parse(additional_columns=ADDITIONAL_COLUMNS, return_gef_data_obj=True)
+
+            # Get the water level from user input, or calculate it from GEF
+            if saved_ground_water_level is not None:
+                ground_water_level = saved_ground_water_level
+            else:
+                ground_water_level = get_water_level(cpt_data_object)
+
+            # Classify the CPTData object to get a SoilLayout
+            soil_layout_obj = cpt_data_object.classify(
+                method=self.method(ground_water_level), return_soil_layout_obj=True
+            )
+
+        except GEFParsingException as e:
+            raise UserException(f"CPT Parsing: {str(e)}")
+        except GEFClassificationError as e:
+            raise UserException(f"CPT Classification: {str(e)}")
+
+        soil_layout_filtered = soil_layout_obj.filter_layers_on_thickness(
+            min_layer_thickness=DEFAULT_MIN_LAYER_THICKNESS,
+            merge_adjacent_same_soil_layers=True,
+        )
+        soil_layout_filtered_in_m = convert_soil_layout_from_mm_to_meter(soil_layout_filtered)
+
+        # Serialize the parsed CPT File content and update it with the new soil layout
+        cpt_dict = cpt_data_object.serialize()
+        cpt_dict["soil_layout_original"] = soil_layout_obj.serialize()
+        cpt_dict["bottom_of_soil_layout_user"] = ceil(soil_layout_obj.bottom) / 1e3
+        cpt_dict["soil_layout"] = convert_soil_layout_to_input_table_field(soil_layout_filtered_in_m)
+        cpt_dict["ground_water_level"] = ground_water_level
+        cpt_dict["x_rd"] = cpt_dict["headers"]["x_y_coordinates"][0] if "x_y_coordinates" in cpt_dict["headers"] else 0
+        cpt_dict["y_rd"] = cpt_dict["headers"]["x_y_coordinates"][1] if "x_y_coordinates" in cpt_dict["headers"] else 0
+        cpt_dict["gef"] = {"cpt_data": {"min_layer_thicknes": DEFAULT_MIN_LAYER_THICKNESS}}
+        return cpt_dict